// src/api/generated/model/productListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Shop API
 * RN 마이그레이션 검증용 API. 에러는 HTTP 상태가 아니라 `code`로 분기한다.
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSummary } from './productSummary';

export interface ProductListResponse {
  items: ProductSummary[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  total: number;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  skip: number;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  limit: number;
}